Add tests for CDATA escape in LegacyXmlResultFormatter
Non-regression test for Bugzilla Report 65833

This also updates JUnit 4 to 4.13.2, JUnit 5 to 5.8.2, and JUnit
Platform to 1.8.2. This also adds junit-platform-testkit as a test
dependency.

Note: At time of this commit, tests for junitlauncher must be run from
the pom.xml for junitlauncher. Attempting to modify build.xml to run the
tests will result in a failure due to the presence of JUnit 3.8.2 on the
classpath.

Signed-off-by: Taylor Smock <tsmock@fb.com>